+++ gazebo/util/LogPlay.hh
@@ -25,6 +25,7 @@
 #include <fstream>
 
 #include "gazebo/common/SingletonT.hh"
+#include "gazebo/common/Time.hh"
 #include "gazebo/util/system.hh"
 
 namespace gazebo
@@ -77,6 +78,14 @@ namespace gazebo
       /// \return The random number seed the open log file. The current
       /// random number seed, as defined in math::Rand::GetSeed.
       public: uint32_t GetRandSeed() const;
+
+      /// \brief Get the log start time of the open log file.
+      /// \return Start time of the log.
+      public: common::Time GetLogStartTime() const;
+
+      /// \brief Get the log end time of the open log file.
+      /// \return End time of the log.
+      public: common::Time GetLogEndTime() const;
 
       /// \brief Step through the open log file.
       /// \param[out] _data Data from next entry in the log file.
@@ -133,6 +142,12 @@ namespace gazebo
 
       /// \brief The random number seed recorded in the open log file.
       private: uint32_t randSeed;
+
+      /// \brief Log start time (simulation time).
+      private: common::Time logStartTime;
+
+      /// \brief Log end time (simulation time).
+      private: common::Time logEndTime;
 
       /// \brief The encoding for the current chunk in the log file.
       private: std::string encoding;